# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: NASA Weather Data API
  description: |
    Weather data and prediction service using NASA POWER API, plus geomagnetic storm data and forecasts.
    
    This API provides historical weather data and future weather predictions based on NASA POWER satellite data.
    For past dates, it returns actual historical measurements. For future dates, it calculates predictions
    using arithmetic mean of historical data from the past 20 years (configurable).
    
    Additionally, the API provides geomagnetic storm data from NASA DONKI and forecasts from NOAA SWPC.
    
    ## Features
    - Historical weather data from NASA POWER API
    - Weather predictions based on historical averages
    - Geomagnetic storm data from NASA DONKI
    - 3-day and 27-day geomagnetic forecasts from NOAA SWPC
    - Multiple weather parameters support
    - Bulk data processing
    - CSV and JSON output formats
    - Comprehensive error handling
    
    ## Data Sources
    - **Weather Data**: NASA's POWER (Prediction Of Worldwide Energy Resources) API - satellite-derived meteorological data
    - **Geomagnetic Storms**: NASA DONKI (Space Weather Database Of Notifications, Knowledge, Information)
    - **Geomagnetic Forecasts**: NOAA Space Weather Prediction Center (SWPC)
  version: 1.0.0
  contact:
    name: NASA Space Apps Challenge Team
    email: support@weatherapi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
  /weather/data:
    get:
      summary: Get weather data for specific location and date
      description: |
        Returns historical weather data for past dates or predictions for future dates based on historical averages.
        
        **Historical Data**: For dates in the past, returns actual measured data from NASA POWER API.
        **Predictions**: For future dates, calculates predictions using arithmetic mean of historical data from the specified number of years.
      operationId: getWeatherData
      tags:
        - Weather Data
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: float
            minimum: -90
            maximum: 90
          example: 41.0
          description: Latitude coordinate in decimal degrees
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
          example: 29.0
          description: Longitude coordinate in decimal degrees
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-10-30"
          description: Date in YYYY-MM-DD format (accepts flexible formats like 2025-1-1, 2025-10-3, or 2025-10-30)
        - name: parameters
          in: query
          required: false
          schema:
            type: string
            pattern: '^[A-Z0-9_]+(,[A-Z0-9_]+)*$'
          example: "T2M,RH2M,WS10M"
          description: Comma-separated list of weather parameters (defaults to T2M,RH2M,WS10M,WD10M,PS,ALLSKY_SFC_SW_DWN)
        - name: historicalYears
          in: query
          required: false
          schema:
            type: integer
            minimum: 5
            maximum: 30
            default: 20
          example: 20
          description: Number of historical years to use for future predictions
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, csv]
            default: json
          example: "json"
          description: Response format
      responses:
        '200':
          description: Successfully retrieved weather data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherDataResponse'
            text/csv:
              schema:
                type: string
                description: CSV formatted weather data
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: External API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: Request timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /weather/historical-range:
    get:
      summary: Get available data range for location
      description: Returns the available date range for historical and prediction data at given coordinates
      operationId: getHistoricalRange
      tags:
        - Weather Data
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: float
            minimum: -90
            maximum: 90
          description: Latitude coordinate in decimal degrees
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
          description: Longitude coordinate in decimal degrees
      responses:
        '200':
          description: Successfully retrieved date range information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DateRangeResponse'
        '400':
          description: Invalid coordinates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /weather/bulk:
    post:
      summary: Process multiple weather data requests
      description: |
        Accepts multiple location-date combinations for batch processing.
        Maximum 10 requests per batch operation.
      operationId: getBulkWeatherData
      tags:
        - Weather Data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkRequest'
      responses:
        '200':
          description: Bulk request processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkResponse'
        '400':
          description: Invalid bulk request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /geomagnetic/storms:
    get:
      summary: Get geomagnetic storm data from NASA DONKI
      description: |
        Returns geomagnetic storm (GST) data from NASA's DONKI (Space Weather Database Of Notifications, Knowledge, Information) system.
        Includes Kp index values, linked space weather events, and storm notifications.
      operationId: getGeomagneticStorms
      tags:
        - Geomagnetic Data
      parameters:
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
          example: "2025-09-05"
          description: Start date for storm data range (YYYY-MM-DD). Defaults to 30 days ago.
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
          example: "2025-10-05"
          description: End date for storm data range (YYYY-MM-DD). Defaults to current date.
      responses:
        '200':
          description: Successfully retrieved geomagnetic storm data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeomagneticStormResponse'
        '400':
          description: Invalid date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid NASA API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: NASA DONKI service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /geomagnetic/forecast/3-day:
    get:
      summary: Get 3-day geomagnetic forecast from NOAA SWPC
      description: |
        Returns the 3-day geomagnetic activity forecast from NOAA Space Weather Prediction Center.
        Includes daily Kp index predictions and activity levels.
      operationId: get3DayForecast
      tags:
        - Geomagnetic Data
      responses:
        '200':
          description: Successfully retrieved 3-day forecast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreeDayForecastResponse'
        '503':
          description: NOAA SWPC service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /geomagnetic/forecast/27-day:
    get:
      summary: Get 27-day geomagnetic outlook from NOAA SWPC
      description: |
        Returns the 27-day geomagnetic activity outlook from NOAA Space Weather Prediction Center.
        Provides extended forecast for potential geomagnetic storms.
      operationId: get27DayOutlook
      tags:
        - Geomagnetic Data
      responses:
        '200':
          description: Successfully retrieved 27-day outlook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwentySevenDayOutlookResponse'
        '503':
          description: NOAA SWPC service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /geomagnetic/forecast/combined:
    get:
      summary: Get combined geomagnetic forecasts (3-day + 27-day)
      description: |
        Returns both 3-day and 27-day geomagnetic forecasts in a single response.
        Useful for getting complete near-term and extended outlook data.
      operationId: getCombinedForecast
      tags:
        - Geomagnetic Data
      responses:
        '200':
          description: Successfully retrieved combined forecast data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CombinedForecastResponse'
        '503':
          description: NOAA SWPC service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /asteroids/feed:
    get:
      summary: Get asteroids by closest approach date
      description: |
        Retrieve a list of Asteroids based on their closest approach date to Earth. 
        Returns detailed information about each asteroid including size estimates, velocity, 
        miss distance, and hazard classification.
        
        **Date Range**: Maximum 7 days between start_date and end_date. 
        If end_date is not provided, it defaults to 7 days after start_date.
      operationId: getAsteroidFeed
      tags:
        - Asteroids
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
            example: "2025-09-07"
          description: Starting date for asteroid search (YYYY-MM-DD format)
        - name: end_date
          in: query
          required: false
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
            example: "2025-09-08"
          description: Ending date for asteroid search (YYYY-MM-DD format). Defaults to 7 days after start_date. Maximum range is 7 days.
      responses:
        '200':
          description: Successfully retrieved asteroid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsteroidFeedResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsteroidErrorResponse'
        '502':
          description: External API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /eonet/events:
    get:
      summary: Get natural disaster events
      description: |
        Quickly get a global snapshot of the most recent major events. Default returns "open" (ongoing) events. 
        
        **Use Case**: Dashboard initial load for "Current Natural Events" list.
        
        **Presentation Logic**: Card list (icon by category), last event timestamp + short magnitude label (e.g., "55 kts").
      operationId: getEonetEvents
      tags:
        - Natural Events
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed]
            default: open
          example: "open"
          description: Event status - "open" for ongoing events or "closed" for ended events
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 5
          description: Maximum number of events to return (default 20, max 100)
      responses:
        '200':
          description: Successfully retrieved natural disaster events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EONETEventsResponse'
  /eonet/events/geojson:
    get:
      summary: Get events in GeoJSON format
      description: |
        Provides the same event data in GeoJSON format for mapping libraries (Leaflet, Google Maps).
        
        **Use Case**: Direct integration with map visualization libraries.
      operationId: getEonetEventsGeoJSON
      tags:
        - Natural Events
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed]
            default: open
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [drought, dustHaze, earthquakes, floods, landslides, manmade, seaLakeIce, severeStorms, snow, tempExtremes, volcanoes, waterColor, wildfires]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successfully retrieved events in GeoJSON format
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    description: GeoJSON FeatureCollection
  /eonet/events/category/{categoryId}:
    get:
      summary: Get category-specific events
      description: |
        Category-focused risk monitoring (e.g., fire smoke / air quality impact).
        
        **Use Case**: Input trigger for air quality or environmental alert module.
        
        **Presentation Logic**: In "Active Wildfires" section, show size (acres) + last update; if data is older than 48 hours, display a "stale data" tag.
      operationId: getEonetEventsByCategory
      tags:
        - Natural Events
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            enum: [drought, dustHaze, earthquakes, floods, landslides, manmade, seaLakeIce, severeStorms, snow, tempExtremes, volcanoes, waterColor, wildfires]
          example: "wildfires"
          description: Category ID (e.g., "wildfires", "severeStorms")
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed]
            default: open
          description: Event status
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 3
          description: Maximum number of events to return
      responses:
        '200':
          description: Successfully retrieved category-specific events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EONETEventsResponse'
        '400':
          description: Invalid category ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EONETErrorResponse'
  /eonet/events/regional:
    get:
      summary: Get regional events with proximity analysis
      description: |
        Detect if there are any events near the user's active location (local risk awareness).
        
        **Use Case**: After user selects location; compute distance and classify "<50 km: High Risk", "50–150 km: Medium".
        
        **Presentation Logic**: Map pin + single-line alert ("Nearby active wildfire"). If event date is old, display in grey.
      operationId: getEonetEventsRegional
      tags:
        - Natural Events
      parameters:
        - name: bbox
          in: query
          required: true
          schema:
            type: string
            pattern: '^-?\d+\.?\d*,-?\d+\.?\d*,-?\d+\.?\d*,-?\d+\.?\d*$'
          example: "26.5,38.0,28.0,39.0"
          description: Bounding box coordinates "minLon,minLat,maxLon,maxLat"
        - name: userLat
          in: query
          required: false
          schema:
            type: number
            format: float
          example: 38.5
          description: User's latitude for distance calculation
        - name: userLon
          in: query
          required: false
          schema:
            type: number
            format: float
          example: 27.0
          description: User's longitude for distance calculation
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, closed]
            default: open
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Successfully retrieved regional events with proximity data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EONETRegionalEventsResponse'
...
  /openaq/measurements/{sensorId}:
    get:
      summary: Get sensor measurements
      description: |
        Get the latest air quality measurement for a specific sensor.
        Returns the most recent measurement with quality assessment.
      operationId: getSensorMeasurements
      tags:
        - Air Quality
      parameters:
        - name: sensorId
          in: path
          required: true
          description: Unique sensor ID from OpenAQ
          schema:
            type: integer
            example: 7017785
      responses:
        '200':
          description: Successfully retrieved sensor measurements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAQMeasurementsResponse'
        '400':
          description: Invalid sensor ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No measurements found for sensor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: OpenAQ API unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    EONETErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_CATEGORY"
            message:
              type: string
              example: "Invalid category ID provided"
            details:
              type: object
        requestTimestamp:
          type: string
          format: date-time
...